+++ pkg/filter/trace/trace.go
@@ -16,12 +16,10 @@ package trace
 
 import (
 	"fmt"
-	"github.com/xmidt-org/ears/internal/pkg/rtsemconv"
 	"github.com/xmidt-org/ears/pkg/event"
 	"github.com/xmidt-org/ears/pkg/filter"
 	"github.com/xmidt-org/ears/pkg/secret"
 	"github.com/xmidt-org/ears/pkg/tenant"
-	"go.opentelemetry.io/otel"
 )
 
 func NewFilter(tid tenant.Id, plugin string, name string, config interface{}, secrets secret.Vault) (*Filter, error) {
@@ -52,12 +50,9 @@ func (f *Filter) Filter(evt event.Event) []event.Event {
 		})
 		return nil
 	}
+
+	//TODO fix this
 	traceId := "123-456-789-000"
-	if evt.Trace() {
-		tracer := otel.Tracer(rtsemconv.EARSTracerName)
-		_, span := tracer.Start(evt.Context(), "traceFilter")
-		defer span.End()
-	}
 	evt.SetPathValue(f.config.Path, traceId, true)
 	return []event.Event{evt}
 }