+++ pkg/plugins/kafka/receiver_config.go
@@ -42,9 +42,6 @@ func (rc *ReceiverConfig) WithDefaults() ReceiverConfig {
 	if cfg.ChannelBufferSize == nil {
 		cfg.ChannelBufferSize = DefaultSenderConfig.ChannelBufferSize
 	}
-	if cfg.Trace == nil {
-		cfg.Trace = DefaultReceiverConfig.Trace
-	}
 	return cfg
 }
 
@@ -103,11 +100,7 @@ const receiverSchema = `
                 },
                 "channelBufferSize": {
                     "type": "integer"
-                },
-				"trace" : {
-					"type": "boolean",
-					"default": false
-				}
+                }
             },
             "required": [
                 "brokers", "topic", "groupId"